requirements/local.txt
# Local development dependencies go here
-r test.txt
django-rest-swagger
django-debug-toolbar
Sphinx